feature/user stock status
Added new feature: store and show user stock total status

+++ controllers/users.js
@@ -20,7 +20,7 @@ class Users {
         type: Sequelize.STRING(20),
       },
       balance: {
-        type: Sequelize.INTEGER(),
+        type: Sequelize.DOUBLE(),
       },
     }, {
       freezeTableName: true,
